+++ src/toad/paths.py
@@ -5,7 +5,7 @@ from typing import Final
 from xdg_base_dirs import xdg_config_home, xdg_data_home, xdg_state_home
 
 
-APP_NAME: Final[str] = "toad"
+APP_NAME: Final[str] = "taiji"
 
 
 def path_to_name(path: Path) -> str: