+++ src/toad/widgets/welcome.py
@@ -4,19 +4,11 @@ from textual import containers
 from textual.widgets import Label, Markdown
 
 
-ASCII_TOAD = r"""
-         _   _
-        (.)_(.)
-     _ (   _   ) _
-    / \/`-----'\/ \
-  __\ ( (     ) ) /__
-  )   /\ \._./ /\   (
-   )_/ /|\   /|\ \_(
-"""
+ASCII_TAIJI = "☯"
 
 
 WELCOME_MD = """\
-## Toad v1.0
+## Taiji v1.0
 
 Welcome, **Will**!
 
@@ -27,5 +19,5 @@ Welcome, **Will**!
 class Welcome(containers.Vertical):
     def compose(self) -> ComposeResult:
         with containers.Center():
-            yield Label(ASCII_TOAD, id="logo")
+            yield Label(ASCII_TAIJI, id="logo")
         yield Markdown(WELCOME_MD, id="message", classes="note")